Major refactor of symbol handling and compilation
Symbol types were changed into bitfields to allow richer descriptions of
symbols in the program. Literal data support was added to symbol tables,
but the storage and transfer of this data is disgusting.

The current priority is to tidy and refactor symbol and syntax tree handling
within the compiler source. The main part of this will be a unionization of
the AST and symbol table literal data values.