Rewrite of compiler error string code
This commit modifies the compiler error string formatting function
in order to maintain consistency in error messages produced by the
earlier compilation phases.

From the way things are going, it is apparent that a proper solution
to the problem of error message formatting is one that never lets
the user know that there is C code running under the hood. The
current error handling system will probably get rewritten, since
errors within the matte runtime will need to be objects anyways...