Contextual errors form horzcat/vertcat
The Parser was modified to build AST row and column nodes with
context information, enabling the inclusion of error messages
for object_horzcat() and object_vertcat().

More simplification of the AST structures related to rows
and columns was also included. Basically anything that can
be compressed down will be now.

There also appears to be a bug in the Range code that counts
the number of elements. Placed it on the unofficial to-do list.