Fully removed old error system
This commit redefines the fail() and throw() macro functions to use
the functionality of Exception in the manner outlined in the notes
for the previous commit.

More work is still needed to properly utilize the new error system.
Unfortunately, it looks like contextual errors are going to die in
this process.

Oh well. :)