Further integration of Exception
As it stands, the core object operation and method dispatch functions
will be responsible for the first throw(). All prior functions may
either return NULL or call error() or fail() to indicate failure.

More refactors of lib/compiler.c are necessary in order to properly
integrate exception handling into the compiled code. For now, the
error handling system is broken, as it only checks for NULL values
from calls, instead of checking for Exception objects.