Added zone deallocation to compiled code
This commit closes the loop on the zone allocation codebase by
adding object_free_all() statements at the end of each scope.

However, this addition has broken the handling of global variables
in functions. The solution will require the creation of a zone
allocator for global variables. Assign statements will then either
assign by-pointer for locals or copy for globals.

It would probably be a good idea to always write a matte_main()
function, and only write a simple main() function for non-memory
compilation targets.