Added a few operators and control flow statements
The matte compiler now outputs 'break', 'continue' and 'return'
statements. This is without any error checking! It would be smart
to move up the syntax tree and make sure that 'break' and 'continue'
are indeed within an enclosing loop.

The language also recognizes element-wise combined assignment
operators like .*= and .^= because why not. Combined operators
are really just parse-time syntactic sugar, so it was trivial to
include the extra support.