Initial support for for loops
One of the most challenging flow control structures in the language
is the for loop, because it accepts many different 'iterable' types
instead of just evaluating the truth of an expression. This commit
introduces the Iter object that wraps any matte object and provides
a uniform interface to for-loop iteration.

At the moment, only ranges and vectors are supported as iterable,
but this will change as time progresses.