Added while and do-until support
This commit rounds out the language support of loop blocks with
'while' loops and 'do-until' loops. More type support was also
added to the Iter object.

It's probably a good point to modify the parsing behavior of the
increment and decrement operators. In the end, these operators
will become parse-time syntactic sugar, just as the combined
assigment operators have.