More Complex object support, minor fixes
This commit adds more functionality into the Complex object source code,
so algebra with Int, Range, Float and Complex objects is now nearly
fully supported. The next objects in the work queue are the Vector
and ComplexVector.

The commit also updates the README file with correct ATLAS compilation
for GCC 5.3.0 and 6.1.0. Minor fixes were also made to the Range and
Float object source code.